requirements.txt
PyQt6>=6.6.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
py2app>=0.28.6
dmgbuild>=1.6.0
